% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfCDS.R
\name{dfCDS}
\alias{dfCDS}
\title{dfCDS}
\usage{
dfCDS(cdsData)
}
\arguments{
\item{cdsData}{A sequencing dataset that contains the coding region start and end sites and predicted gene (.tsv)}
}
\value{
A dataframe of coding region genes
}
\description{
Load in CDS data, .tsv
}
