% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfIPD.R
\name{dfIPD}
\alias{dfIPD}
\title{Read IPD csv file}
\usage{
dfIPD(ipdData)
}
\arguments{
\item{ipd_data}{A sequencing kinetics dataset from PacBio SMRT Sequencing kineticsTools pkg, contains base call and ipdRatio (.csv)}
}
\value{
The read file
}
\description{
Load in PacBio methylation data, modifications.csv
}
