% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfMotif.R
\name{dfMotif}
\alias{dfMotif}
\title{dfMotif}
\usage{
dfMotif(motifData)
}
\arguments{
\item{motifData}{A sequencing kinetics dataset that contains the kineticstools detected motif positions (.gff)}
}
\value{
A dataframe of methylation motifs
}
\description{
Load in PacBio motif data, motifs.gff
}
