% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfTFBS.R
\name{dfTFBS}
\alias{dfTFBS}
\title{Read ChipSeq data}
\usage{
dfTFBS(chipseqData)
}
\arguments{
\item{chipseqData}{A chipSeq dataset that contains the regulator (TF), peak position, and score (.gff)}
}
\value{
The read file
}
\description{
Load in ChipSeq data
}
