% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfTranscription.R
\name{dfTranscription}
\alias{dfTranscription}
\title{Read transcriptional data}
\usage{
dfTranscription(transcriptionData)
}
\arguments{
\item{transcriptionData}{A transcriptional dataset that contains the transcript, and relative abundance (.csv)}
}
\value{
The read file
}
\description{
Load in transcriptomics data
}
