% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinmotifGene.R
\name{joinmotifGene}
\alias{joinmotifGene}
\title{joinmotifGene}
\usage{
joinmotifGene(dfCDS, dfmotifIPD)
}
\arguments{
\item{dfCDS, }{A dataframe with CDS (gene) name, start site, and end site}

\item{dfmotifIPD, }{A dataframe with ipdRatio data by position}
}
\value{
A dataframe with combined PacBio methylation data, by motif
}
\description{
Join motifs with ipdRatio call using position and strand
}
