% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinmotifgeneTranscription.R
\name{joinmotifgeneTranscription}
\alias{joinmotifgeneTranscription}
\title{Transcriptional Reponse}
\usage{
joinmotifgeneTranscription(dfmethbasestartGene, dfTranscription)
}
\arguments{
\item{dfmethbasestartGene, }{A motif dataset with ipdRatio and nearest gene}

\item{dfTranscription, }{A transcriptional dataset with gene, and relative abundance of associated mRNA}
}
\value{
The read file
}
\description{
Join methylated base position and associated gene with transcript abundance data, start and end sites
}
