% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinmotifIPD.R
\name{joinmotifIPD}
\alias{joinmotifIPD}
\title{joinmotifIPD}
\usage{
joinmotifIPD(dfIPD, dfMotif)
}
\arguments{
\item{dfIPD}{A dataframe with ipdRatio data by position}

\item{dfMotif}{A dataframe where each row is an identified methylation motif}
}
\value{
A dataframe of combined PacBio sequencing kinetics values by motif site
}
\description{
Join motifs with ipdRatio call using position and strand
}
