% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Motifs}
\alias{Motifs}
\alias{Motif}
\alias{IPD}
\alias{CDS}
\alias{TFBS}
\alias{Transcription}
\title{Motifs of E. coli}
\format{
A data frame with 147336 rows and 9 variables

A data frame with 147336 rows and 9 variables

A data frame with 147336 rows and 9 variables

A data frame with 147336 rows and 9 variables

A data frame with 147336 rows and 9 variables
}
\usage{
data(Motif)

data(IPD)

data(CDS)

data(TFBS)

data(Transcription)
}
\description{
Dataset containing SMRT output of sequencing run

Dataset containing SMRT output of sequencing run

Dataset containing annotated gene locations

Dataset containing chip-seq data of TF binding site locations and strength

Dataset containing chip-seq data of TF binding site locations and strength
}
\keyword{datasets}
