% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfCDS.R
\name{dfCDS}
\alias{dfCDS}
\title{Read CDS}
\usage{
dfCDS(cdsData)
}
\arguments{
\item{cdsData}{A sequencing dataset that contains the coding region start and end sites annotated onto the genome (.tsv)}
}
\value{
The read file
}
\description{
Load in CDS data
}
