% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfIPD.R
\name{dfIPD}
\alias{dfIPD}
\title{dfIPD}
\usage{
dfIPD(ipdData)
}
\arguments{
\item{ipdData}{A sequencing kinetics dataset from PacBio SMRT Sequencing kineticsTools pkg, contains base call and ipdRatio (.csv)}
}
\value{
A dataframe of the genome, that includes the called base and ipdRatio at every base
}
\description{
Load in PacBio methylation data, modifications.csv
}
