% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfMotif.R
\name{dfMotif}
\alias{dfMotif}
\title{Read motif gff file}
\usage{
dfMotif(motifData)
}
\arguments{
\item{motifData}{A sequencing kinetics dataset that contains only detected motif positions (.gff)}
}
\value{
The read file
}
\description{
Load in PacBio motif data, motifs.gff
}
