% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfTFBS.R
\name{dfTFBS}
\alias{dfTFBS}
\title{dfTFBS}
\usage{
dfTFBS(chipseqData)
}
\arguments{
\item{chipseqData}{A ChIPSeq dataset that contains the regulator (TF), peak position, and score (.gff or .csv)}
}
\value{
A dataframe of the TF binding data
}
\description{
Load in ChIPSeq data, .csv or .gff
}
