% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinmotifGene.R
\name{joinmotifGene}
\alias{joinmotifGene}
\title{Create nearest CDS column}
\usage{
joinmotifGene(dfCDS, dfmotifIPD)
}
\arguments{
\item{dfCDS, }{A df with CDS name, start site, and end site}

\item{dfmotifIPD, }{A df with ipdRatio data by position}
}
\value{
Dataframe with combined PacBio methylation data, by motif.
}
\description{
Join motifs with ipdRatio call using position, strand
}
