% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinmotifgeneTranscription.R
\name{joinmotifgeneTranscription}
\alias{joinmotifgeneTranscription}
\title{joinmotifgeneTranscription}
\usage{
joinmotifgeneTranscription(dfmethbasestartGene, dfTranscription)
}
\arguments{
\item{dfmethbasestartGene, }{A methylation motif dataframe with ipdRatio and nearest gene}

\item{dfTranscription, }{A transcriptional dataframe with gene, and relative abundance of associated mRNA}
}
\value{
A dataframe of methylation motifs with each associated gene's transcriptional data
}
\description{
Join methylation motifs and associated gene with transcript abundance data
}
