% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinmotifIPD.R
\name{joinmotifIPD}
\alias{joinmotifIPD}
\title{Create data frame of motifs with associated ipdRatio data}
\usage{
joinmotifIPD(dfIPD, dfMotif)
}
\arguments{
\item{dfIPD, }{A df (data frame) with ipdRatio data by position}

\item{dfMotif, }{A df where each row is an identified methylation motif}
}
\value{
The read file
}
\description{
Join motifs with ipdRatio call using position, strand
}
